Cleanup and dedup. * Use NewValueAge better while listing apps * Relocate packageInstallStatus()
