+++ cli/pkg/kctrl/cmd/package/installed/cmd.go
@@ -5,7 +5,6 @@ package installed
 
 import (
 	"github.com/spf13/cobra"
-	kcpkgv1alpha1 "github.com/vmware-tanzu/carvel-kapp-controller/pkg/apis/packaging/v1alpha1"
 )
 
 func NewCmd() *cobra.Command {
@@ -15,14 +14,4 @@ func NewCmd() *cobra.Command {
 		Short:   "Manage installed packages",
 	}
 	return cmd
-}
-
-func packageInstallStatus(pkgi *kcpkgv1alpha1.PackageInstall) string {
-	if pkgi.Spec.Canceled {
-		return "Canceled"
-	}
-	if pkgi.Spec.Paused {
-		return "Paused"
-	}
-	return pkgi.Status.FriendlyDescription
 }
